flaskProject2/requirements.txt
Flask~=3.0.1
connexion~=3.0.5
requests~=2.31.0
pillow~=10.2.0
marshmallow~=3.20.2